[xmp] Show mod time instead of elapsed time
When replaying, display a consistent mod-event time instead of wallclock
elapsed time.

+++ src/common.h
@@ -29,7 +29,7 @@ struct options {
 };
 
 struct control {
-	double time;		/* Replay time in ms */
+	long time;		/* Replay time in ms */
 	int skip;		/* Skip to next module */
 	int loop;		/* Module is looped */
 	int pause;		/* Replay paused */
